added comments and liquibase

+++ src/main/java/ru/otus/library/dao/AuthorDao.java
@@ -7,11 +7,11 @@ import java.util.Optional;
 
 public interface AuthorDao {
 
-    int count();
+    long count();
     void insert(Author author);
-    Optional<Author> getById(int id);
+    Optional<Author> getById(long id);
     List<Author> getAll();
     void delete(Author author);
     List<Author> findByName(String name);
-    List<Author> findByBookId(int id);
+    List<Author> findByBookId(long id);
 }
