+++ src/main/java/ru/otus/library/services/impl/BookServiceImpl.java
@@ -69,7 +69,7 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Optional<Book> findById(int id) {
+    public Optional<Book> findById(long id) {
         try {
             return bookDao.getById(id);
         } catch (EmptyResultDataAccessException e) {
@@ -80,7 +80,7 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public int delete(Book book) {
+    public long delete(Book book) {
          bookDao.delete(book);
          return book.getId();
     }
